# Conflicts:
#	app/Http/Controllers/API/StatusController.php
#	app/Http/Controllers/API/v1/StatusController.php

+++ app/Http/Controllers/Backend/Transport/TrainCheckinController.php
@@ -8,10 +8,12 @@ use App\Enum\StatusVisibility;
 use App\Events\UserCheckedIn;
 use App\Exceptions\Checkin\AlreadyCheckedInException;
 use App\Exceptions\CheckInCollisionException;
+use App\Exceptions\HafasException;
 use App\Exceptions\NotConnectedException;
 use App\Exceptions\StationNotOnTripException;
 use App\Http\Controllers\Backend\GeoController;
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\HafasController;
 use App\Http\Controllers\StatusController as StatusBackend;
 use App\Http\Controllers\TransportController;
 use App\Http\Resources\PointsCalculationResource;
@@ -67,15 +69,15 @@ abstract class TrainCheckinController extends Controller
             throw new InvalidArgumentException('Departure time must be before arrival time');
         }
 
-        $status = StatusBackend::createStatus(
-            user:       $user,
-            business:   $travelReason,
-            visibility: $visibility,
-            body:       $body,
-            eventId:    $event?->id
-        );
-
         try {
+            $status = StatusBackend::createStatus(
+                user:       $user,
+                business:   $travelReason,
+                visibility: $visibility,
+                body:       $body,
+                eventId:    $event?->id
+            );
+
             $trainCheckinResponse = self::createTrainCheckin(
                 status:      $status,
                 trip:        $hafasTrip,
@@ -93,9 +95,19 @@ abstract class TrainCheckinController extends Controller
             );
 
             return $trainCheckinResponse;
+        } catch (PDOException $exception) {
+            if (isset($status)) {
+                $status->delete();
+            }
+            if ((int) $exception->getCode() === 23000) { // Integrity constraint violation: Duplicate entry
+                throw new AlreadyCheckedInException();
+            }
+            throw $exception; // Other scenarios are not handled
         } catch (Exception $exception) {
             // Delete status if it was created and rethrow exception, so it can be handled by the caller
-            $status->delete();
+            if (isset($status)) {
+                $status->delete();
+            }
             throw $exception;
         }
     }
@@ -214,10 +226,43 @@ abstract class TrainCheckinController extends Controller
         $checkin->update([
                              'arrival'     => $newDestinationStopover->arrival_planned->toIso8601String(),
                              'destination' => $newDestinationStopover->trainStation->ibnr,
+                             'distance'    => $newDistance,
                              'points'      => $pointsResource['points'],
                          ]);
+        $checkin->refresh();
 
         return PointReason::tryFrom($pointsResource['calculation']['reason']);
     }
 
+    /**
+     * @param string $tripId
+     * @param string $lineName
+     * @param int    $start
+     *
+     * @return HafasTrip
+     * @throws HafasException
+     * @throws StationNotOnTripException
+     * @api v1
+     */
+    public static function getHafasTrip(string $tripId, string $lineName, int $startId): HafasTrip {
+        $hafasTrip = HafasController::getHafasTrip($tripId, $lineName);
+        $hafasTrip->loadMissing(['stopoversNEW', 'originStation', 'destinationStation']);
+
+        $originStopover = $hafasTrip->stopoversNEW->filter(function(TrainStopover $stopover) use ($startId) {
+            return $stopover->train_station_id === $startId || $stopover->trainStation->ibnr === $startId;
+        })->first();
+
+        if ($originStopover === null) {
+            throw new StationNotOnTripException();
+        }
+
+        //try to refresh the departure time of the origin station
+        if ($originStopover) {
+            dispatch(function() use ($originStopover) {
+                HafasController::refreshStopover($originStopover);
+            })->afterResponse();
+        }
+
+        return $hafasTrip;
+    }
 }
